backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.0
requests==2.31.0
openai==1.3.0
groq==0.4.2
google-generativeai==0.3.2
python-dotenv==1.0.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
Werkzeug==3.0.1
